refactor(FON-468): group espace shells under pages/spaces
