+++ apps/client/src/pages/spaces/admin/jobs/DetailsJobPage.tsx
