+++ apps/client/src/pages/spaces/admin/jobs/common/jobs-page-outlet-context.type.ts
