+++ apps/client/src/pages/spaces/admin/jobs/components/JobFileNode.tsx
