+++ apps/client/src/pages/spaces/admin/jobs/components/JobsListItem.tsx
