+++ apps/client/src/pages/spaces/member/MemberLayout.tsx
@@ -1,8 +1,8 @@
 import { Outlet, useLocation, useParams } from 'react-router';
 
-import { AuthGuard } from '../../components/guards/AuthGuard';
-import { PageContentLayout } from '../../components/shared/PageContentLayout';
-import { AUTHORIZED_ROLES } from '../../constants/authorized-roles.constants';
+import { AuthGuard } from '../../../components/guards/AuthGuard';
+import { PageContentLayout } from '../../../components/shared/PageContentLayout';
+import { AUTHORIZED_ROLES } from '../../../constants/authorized-roles.constants';
 
 export const MemberLayout = () => {
   const params = useParams();
